requirements.txt
# Binance API
python-binance==1.0.19
ccxt==4.2.0

# Async
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Redis
redis==5.0.1
hiredis==2.3.2

# Data processing
pandas==2.1.4
numpy==1.26.2
ta==0.11.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.2

# HTTP server (for Railway health checks)
fastapi==0.109.0
uvicorn==0.25.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
